ML pages added with features 3
